Finished section 1-2
In the BP_BaseWeapon class added a function to implement a line trace system. From all weapons made from this base class, when they are shot there will be a ray shot in the direction the player is facing, and it will check for collisions with the world and enemies. Also started creating the pistol weapon. Added a custom event that will shoot the weapon, play its animation, and call the line trace function from the base class. Have yet to implement controls for the pistol.