Finished section 1-4
Created an ammo system for 4 different ammo types: bullets, shells, rockets, cells. Using an enumerator, within the weapon base class created a function called HasEnoughAmmo?, which checks if a particular ammo type is greater than 0. Each weapon that is a child of the base weapon class will be able to set its weapon type using the enumerator. Then in the event graph created an event called DecreaseAmmo, which decrements the integer value of the ammo type. These two new functions are called when the player presses the shoot button.