Finished section 1-6
Added function to the PlayerHUD that updates the values of the ammo for all the different ammo types with the amount of ammo the player has, as well as the maximum amount per ammo type. This is done on startup. Then in the BaseWeapon class, in its decrease ammo function, for each ammo type the update (insert ammo type) ammo function from the hud is called, in order to update the hud each time the player shoots.