Finished section 2-2
Created the minigun weapon as a child of the base weapon class. Blueprints a bit different from the pistol and shotgun.  When player shoots there is a timer that starts, once that timer ends thats when the minigun actually shoots, and if the player hold the shoot button the timer resets and the minigun shoots again. Once the player lets go of the shoot button, that timer is invalidated through the stop shooting function in the base weapon class. In the first person character class,  once the input for shooting is let go, the return value of the event becomes completed, which is connected to the stop shooting function.