Finished section 2-4
Finished making the plasma gun and projectile system. Added a function in the base weapon class called FireProjectile. This function is able to spawn a projectile with a given speed and damage and transform (its position in the world and the direction it is meant to be facing). This function is then called in the plasma gun which is identical to the minigun except instead of using the fire weapon function which creates a ray, it calls this FireProjectile function.