Finished section 2-6
Created the BFG weapon and projectile, its shoot function works the same as the other projectile weapons thus far, except it is checking that it has at least 30 ammo in the cell ammo category in order to shoot. However, there is a bug where even when there is exactly 30 ammo left, it is considering it as not enough ammo.