Finished section 2-7
Fixed 2 bugs. The first bug was how the BFG was not firing even when it was on its last 30 cell ammo, by changing the conditional from > 30 cell ammo to >= 30 cell ammo. The second bug: the initial goal for the weapons, excluding the minigun and plasma gun, is that the player would need to click the shoot button every single time they wanted to shoot, and the minigun/plasma gun were the only weapons where they would be able to hold down the shoot button and continuously fire. This was fixed through the first person character class blueprint. Stemming from the Input Action Fire (in this case left click), the shoot weapon function was triggered. However the bug was happening because it stemmed from the "triggered" portion of the input action fire. As in, if the left click was held down, it would keep sending the signal, thus the player was able to hold down the shoot button for weapons where that was not meant to be the case. Instead it now stems from the "started" portion of the input action fire.