Finished section 2-10
Started creating a weapon pickup system. In the first person character class created a event called pickupweapon, which changes the child actor(the weapon), to whatever the weapon being picked up is. Then created a base class for weapon pickups which contains two event. The first, when the weapon pickup is collided with, it calls the pickup weapon event from the first person character class. Then the second being a event where every tick, the sprite of the weapon pickup faces the character by manipulating its z rotation.