Finished section 2-11
Added on to the weapon pickup system. Within the weapon pickup class, added to the construction script (Script that will run once the object is instanced), and using an enumerator with all the different weapon types, set the sprite of the weapon pickup, as well as the base weapon class object depending on what we want the weapon to be. This will also impact the player function for picking up the weapon, as now this enumerator construction script now dictates the weapon class variable that will be picked up.