Finished section 2-12
Started creating a weapon bob system. So far made weapon bob from left to right. This was done using a timeline of 2 seconds, which goes from the value of 0-1, then 1-0. Then changed the relative y position of the weapon sprite using the lerp function. When the value of the timeline hits a 0, it starts increasing the y position by 0.02, then once it hits a 1, it decreases it by 0.02. This enables the weapon sprite to slowly move from left to right.