Finished section 2-18
Added text to the HUD showing the current ammo being used. It is its own variable in the HUD class and the text gets updated using a new function, where depending on the numerator value of "weapon" it updates the text to that specific value. This function is used in the swapping weapons event and the pickup weapons event.