Finished section 2-4
Added a frame by frame system to check which direction the player is relative to the enemy. First by grabbing a normalized vector of the player's position and the enemy's subtracted. Then taking the dot product of that vector and the enemy's forward vector. Use arccos on that value to get the degrees between the forward vector and the player. However, this number on reaches 180 then loops back around to the other side of the enemy, thus we need a way to determine whether or not the player is on the left or the right side of the enemy. To do this, we then take the dot product of the normalized vector from before, and the enemy's right vector, then check if it is less than 0 (a negative number). If negative, that means we are on the left side of the enemy, otherwise we are on the right.