Finished section 3-5
Created an 8 directional sprite system for the enemy. Depending on the degrees at which the player is away from the front vector of the enemy, it will rotate the sprite to that its facing the player as well as change the sprite itself to reflect which part of the enemy the player is seeing.